NN-1663 Update existing tests
