+++ app/containers/Bookings/Details/Scheduled/elements/DayWithDate.js
@@ -1,8 +1,8 @@
-import React from 'react'
+import React, { Fragment } from 'react'
 import PropTypes from 'prop-types'
 import moment from 'moment'
 
-const DayWithDate = ({ value }) => <span>{moment(value).format('dddd')}</span>
+const DayWithDate = ({ value }) => <Fragment>{moment(value).format('dddd')}</Fragment>
 
 DayWithDate.propTypes = {
   value: PropTypes.string.isRequired,
