+++ app/containers/Bookings/Details/Scheduled/elements/Event.js
@@ -3,39 +3,36 @@ import PropTypes from 'prop-types'
 import moment from 'moment'
 import styled from 'react-emotion'
 import { SPACING } from '@govuk-react/constants'
+import { typography } from '@govuk-react/lib'
 
 export const StyledEvent = styled('div')`
   margin-bottom: ${SPACING.SCALE_3};
-  /* font-size: 16px; */
+  ${typography.font({ size: 19 })};
 
   :last-of-type {
     margin-bottom: 0;
   }
+
+  @media print {
+    font-size: 14px;
+  }
 `
 
-export const EventTimes = styled('div')``
 export const EventType = styled('div')`
   font-weight: 700;
 `
-export const EventComment = styled('div')``
 
 const Event = ({ startTime, endTime, type, shortComment, cancelled }) => (
   <StyledEvent>
-    <EventTimes>
-      <span className="whereabouts-startTime">{moment(startTime).format('HH:mm')}</span>
-
-      {endTime && (
-        <Fragment>
-          {' '}
-          - <span className="whereabouts-endTime"> {moment(endTime).format('HH:mm')} </span>
-        </Fragment>
-      )}
-    </EventTimes>
-
+    <span className="whereabouts-startTime">{moment(startTime).format('HH:mm')}</span>
+    {endTime && (
+      <Fragment>
+        {' '}
+        - <span className="whereabouts-endTime"> {moment(endTime).format('HH:mm')} </span>
+      </Fragment>
+    )}
     <EventType>{type}</EventType>
-
-    <EventComment>{shortComment}</EventComment>
-
+    {shortComment}
     {cancelled && <span className="cancelled"> (cancelled)</span>}
   </StyledEvent>
 )
