+++ app/containers/Bookings/Details/Scheduled/elements/Event.test.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 
-import { Event } from '../index'
+import Event from './Event'
 
 describe('Event component', () => {
   it('should format the startTime and endTime into a time only format', () => {
     const component = shallow(
-      <Event startTime="2017-12-12T08:00" endTime="2017-12-12T22:00" cancelled="no" shortComment="" type="" />
+      <Event startTime="2017-12-12T08:00" endTime="2017-12-12T22:00" cancelled={false} type="" />
     )
     const startTime = component.find('.whereabouts-startTime')
     const endTime = component.find('.whereabouts-endTime')
@@ -16,7 +16,7 @@ describe('Event component', () => {
   })
 
   it('should handle the scenario where no endTime supplied', () => {
-    const component = shallow(<Event startTime="2017-12-12T08:00" endTime="" cancelled="" shortComment="" type="" />)
+    const component = shallow(<Event startTime="2017-12-12T08:00" cancelled={false} type="" />)
     const endTime = component.find('.whereabouts-endTime')
 
     expect(endTime.length).toBe(0)
